# Use LLVM ADTs to improve compile times