--- a/compiler/include/scopeResolve.h
+++ b/compiler/include/scopeResolve.h
@@ -21,6 +21,9 @@
 #ifndef _SCOPE_RESOLVE_H_
 #define _SCOPE_RESOLVE_H_
 
+#include "llvm/ADT/SmallVector.h"
+
+class astlocT;
 class BaseAST;
 class CallExpr;
 class DefExpr;
@@ -42,7 +45,7 @@ Symbol*  lookup(const char*           name,
 
 void     lookup(const char*           name,
                 BaseAST*              context,
-                std::vector<Symbol*>& symbols,
+                llvm::SmallVectorImpl<Symbol*>& symbols,
                 std::map<Symbol*, astlocT*>& renameLocs,
                 std::map<Symbol*, VisibilityStmt*>& reexportPts,
                 bool storeRenames = false);
@@ -60,7 +63,7 @@ Symbol*  lookupAndCount(const char*           name,
 Symbol* lookupInModuleOrBuiltins(ModuleSymbol* mod, const char* name,
                                 int& nSymbolsFound);
 
-void checkConflictingSymbols(std::vector<Symbol *>& symbols,
+void checkConflictingSymbols(llvm::SmallVectorImpl<Symbol *>& symbols,
                              const char* name,
                              BaseAST* context,
                              bool storeRenames,