--- a/compiler/optimizations/deadCodeElimination.cpp
+++ b/compiler/optimizations/deadCodeElimination.cpp
@@ -87,12 +87,13 @@ static bool isDeadVariable(Symbol* var) {
 }
 
 void deadVariableElimination(FnSymbol* fn) {
-  std::set<Symbol*> symSet;
+  llvm::SmallPtrSet<Symbol*, 32> symSet;
   collectSymbolSet(fn, symSet);
 
   // Use 'symSet' and 'todo' together for a unique queue of symbols to process
   std::queue<Symbol*> todo;
-  for_set(Symbol, sym, symSet) {
+  // for_set(Symbol, sym, symSet) {
+  for(Symbol* sym : symSet) {
     todo.push(sym);
   }