--- a/compiler/passes/normalize.cpp
+++ b/compiler/passes/normalize.cpp
@@ -1419,7 +1419,7 @@ void addMentionToEndOfStatement(Expr* node, CallExpr* existingEndOfStatement) {
     return;
 
   // Gather symexprs used in the statement
-  std::vector<SymExpr*> mentions;
+  llvm::SmallVector<SymExpr*, 32> mentions;
   collectSymExprs(node, mentions);
 
   SET_LINENO(node);
@@ -1444,7 +1444,7 @@ void addMentionToEndOfStatement(Expr* node, CallExpr* existingEndOfStatement) {
   // the variable lifetime still matches the user's view of the code.
   // A reasonable alternative would be for transformations such as
   // the removal of .type blocks to add such SymExprs.
-  for_vector(SymExpr, se, mentions) {
+  for (SymExpr* se : mentions) {
     if (VarSymbol* var = toVarSymbol(se->symbol())) {
       if (!var->hasFlag(FLAG_TEMP) &&
           !var->isParameter() &&