--- a/compiler/resolution/resolveFunction.cpp
+++ b/compiler/resolution/resolveFunction.cpp
@@ -48,6 +48,7 @@
 #include "TryStmt.h"
 #include "view.h"
 #include "WhileStmt.h"
+#include "llvm/ADT/DenseSet.h"
 
 #include <set>