… compiler/resolution/initializerResolution.cpp
… compiler/resolution/visibleFunctions.cpp
… compiler/resolution/wrappers.cpp