This PR adjusts the compiler to use LLVM ADTs from the LLVMSupport library. This PR depends on #19565 to build LLVMSupport even with CHPL_LLVM=none. This provides an approximately 6% performance improvement to the performance of compiling Hello World. 

Changes here were authored by @aconsroe-hpe in https://github.com/Cray/chapel-private/issues/2834

Hello World C backend compile times on my home system - built with quickstart + CHPL_LLVM=none (which includes the LLVM support library after #19565) - showing times from three runs
 * main 195a37574cd5a748cbbc9156bbe732bcc6d63eb4 5.1s 5.1s 5.2s
 * this PR 7f6272e2de525cf8f701f080376f8eac8794c384 4.8s 4.8s 4.9s

And with CHPL_LLVM=system (so using the LLVM backend):
 * main 195a37574cd5a748cbbc9156bbe732bcc6d63eb4 4.8s 4.7s 4.7s
 * this PR 7f6272e2de525cf8f701f080376f8eac8794c384 4.4s 4.4s 4.3s

Reviewed by @benharsh - thanks!

- [x] full local testing