@mppf 2022-07-12
Hello World compile times - built with quickstart + CHPL_LLVM=system - showing times from three runs
 * main 0432111a06 4.48 / 4.54 / 4.47 s
 * use LLVM ADTs 2ad5e45962 4.15 / 4.20 / 4.15 s
 * DenseSet in resolveSignatureAndFunction 9bcfd6be479 4.14 / 4.16 / 4.11 s